Integer to Long quantidade

+++ src/main/java/br/com/clinipet/ClinipetControl/model/entity/ItemVenda.java
@@ -26,7 +26,7 @@ public class ItemVenda {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Integer quantidade;
+    private Long quantidade;
 
     @ManyToOne
     @JoinColumn(name = "idAgendamento")
